% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abreviaNomeTitulos.R
\name{abreviaNomeTitulos}
\alias{abreviaNomeTitulos}
\title{abreviaNomeTitulos}
\usage{
abreviaNomeTitulos(titulosPublicos)
}
\arguments{
\item{titulosPublicos:}{array com os nomes dos títulos Públicos}
}
\description{
os dados vem da seguinte url: http://www.tesourotransparente.gov.br/ckan/dataset/taxas-dos-titulos-ofertados-pelo-tesouro-direto

Retorna array com nomes dos títulos públicos abreviados
}
\author{
Filipe Marques
}
\keyword{abertos,}
\keyword{dados}
\keyword{direto,}
\keyword{publicos}
\keyword{tesouro}
\keyword{titulos}

