% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existsData.R
\name{existsData}
\alias{existsData}
\title{exists data}
\usage{
existsData()
}
\value{
TRUE if
}
\description{
Funcao para verificar se já existe localmente os dados do tesouro direto que estao na pasta data com o nome 'precotaxatesourodireto.csv'
}
\author{
Filipe Marques
}

