% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existsDataCompraTitulo.R
\name{existsDataCompraTitulo}
\alias{existsDataCompraTitulo}
\title{exists data}
\usage{
existsDataCompraTitulo()
}
\value{
TRUE se já existe localmente os dados do Tesouro Direto, FALSE caso contrário
}
\description{
Funcao para verificar se já existe localmente os dados do Tesouro Direto que estão na pasta data com o nome 'comprataxatesourodireto.csv'
}
\author{
Rodrigo Almeida
}

