% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existsDataSelic.R
\name{existsDataSelic}
\alias{existsDataSelic}
\title{exists data}
\usage{
existsDataSelic()
}
\value{
TRUE se já existe localmente os dados do Tesouro Direto, FALSE caso contrário
}
\description{
Funcao para verificar se já existe localmente os dados do Tesouro Direto que estão na pasta data com o nome 'selic.csv'
}
\author{
Rodrigo Almeida
}

