% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existsDataVendaTitulo.R
\name{existsDataVendaTitulo}
\alias{existsDataVendaTitulo}
\title{exists data}
\usage{
existsDataVendaTitulo()
}
\value{
TRUE se já existe localmente os dados do Tesouro Direto, FALSE caso contrário
}
\description{
Funcao para verificar se já existe localmente os dados do Tesouro Direto que estão na pasta data com o nome 'vendataxatesourodireto.csv'
}
\author{
Rodrigo Almeida
}

