% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAliquotaIR.R
\name{getAliquotaIR}
\alias{getAliquotaIR}
\title{get Aliquota do Imposto de Renda}
\usage{
getAliquotaIR(diasUteis)
}
\arguments{
\item{:}{diasUteis: quantidade de dias uteis}
}
\description{
Funcao para recuperar aliquota IR
}
\author{
Rodrigo Almeida (rodrigo.almeida@gmail.com)
}

