% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataSelic.R
\name{getDataSelic}
\alias{getDataSelic}
\title{get data Selic}
\usage{
getDataSelic(updateDataSeNaoDisponivel = FALSE)
}
\arguments{
\item{updateDataSeNaoDisponivel:}{TRUE se deseja fazer download do arquivo de dados caso o mesmo não esteja disponível, FALSE caso o arquivo já esteja disponível.}
}
\description{
Funcao para recuperar dados Selic
}
\author{
Rodrigo Almeida
}

