% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGrafico.R
\name{getGrafico}
\alias{getGrafico}
\title{getGrafico}
\usage{
getGrafico(tituloPublico, dataVencimento, janelaLonga, janelaCurta, tipoInfo,
  dataLimiteInferiorEixoX = "2015-01-01",
  dataLimiteSuperiorEixoX = "2016-11-15", MMECurta, MMELonga)
}
\arguments{
\item{janelaLonga}{de avaliação da media movel exponencial}

\item{janelaCurta}{de avaliação da media movel exponencial}

\item{tituloPublico:}{Titulo Publico}

\item{dataVencimento:}{Data de vencimento do titulo (formato format = "%Y-%m-%d")}

\item{tipoInfo:}{Um dos seguintes: Taxa.Compra.Manha, Taxa.Venda.Manha, PU.Compra.Manha, PU.Venda.Manha, PU.Base.Manha}

\item{dataLimiteInferiorEixoX:}{data Limite Inferior a ser apresentada no Eixo X (formato format = "%Y-%m-%d")}

\item{dataLimiteSuperiorEixoX:}{data Limite Superior a ser apresentada no Eixo X (formato format = "%Y-%m-%d")}

\item{MME12:}{Apresentar Media Movel de 12 dias}

\item{MME26:}{Apresentar Media Movel de 26 dias}
}
\description{
Funcao para geracao de grafico do Tesouro
}
\author{
Rodrigo Almeida
}

