% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraficoComparativoCompraVenda.R
\name{getGraficoComparativoCompraVenda}
\alias{getGraficoComparativoCompraVenda}
\title{getGraficoComparativoCompraVenda}
\usage{
getGraficoComparativoCompraVenda(tituloPublico, dataVencimento)
}
\arguments{
\item{tituloPublico:}{Titulo Publico}

\item{dataVencimento:}{Data de vencimento do titulo (formato format = "%Y-%m-%d")}
}
\description{
Funcao para geracao de grafico Comparando Volumes de Compra e Venda de títulos do Tesouro
}
\author{
Rodrigo Almeida
}

