% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraficoSelic.R
\name{getGraficoSelic}
\alias{getGraficoSelic}
\title{getGraficoComparativoCompraVenda}
\usage{
getGraficoSelic()
}
\description{
Funcao para geracao de grafico Comparando Volumes de Compra e Venda de títulos do Tesouro
}
\author{
Rodrigo Almeida
}

