% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRentabilidadeLTN.R
\name{getRentabilidadeLTN}
\alias{getRentabilidadeLTN}
\title{get Valor Rentabilidade Tesouro Prefixado (LTN)}
\usage{
getRentabilidadeLTN(dataCompra, tituloPublico, dataVencimento, valorInvestido,
  txAdmBanco)
}
\arguments{
\item{:}{dataCompra: Data da Compra do titulo no formato YYYY-MM-DD}

\item{:}{tituloPublico: Nome do Titulo Publico}

\item{:}{dataVencimento: Data de Vencimento do titulo no formato YYYY-MM-DD}

\item{:}{valorInvestido: Valor total investido no periodo}

\item{:}{txAdmBanco: taxa de administracao do banco ou financeira para o periodo (%a.a)}
}
\description{
Funcao para calcular previsão de rentabilidade de titulo LTN
}
\author{
Rodrigo Almeida - rodrigo.almeida@gmail.com
}

