% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRentabilidadeNTNBPrincipal.R
\name{getRentabilidadeNTNBPrincipal}
\alias{getRentabilidadeNTNBPrincipal}
\title{get Valor Rentabilidade Tesouro IPCA  (NTN-B Principal)}
\usage{
getRentabilidadeNTNBPrincipal(dataCompra, tituloPublico, dataVencimento,
  valorInvestido, txAdmBanco, txInflacaoIPCA)
}
\arguments{
\item{:}{dataCompra: Data da Compra do titulo no formato YYYY-MM-DD}

\item{:}{tituloPublico: Nome do Titulo Publico}

\item{:}{dataVencimento: Data de Vencimento do titulo no formato YYYY-MM-DD}

\item{:}{valorInvestido: Valor total investido no periodo}

\item{:}{txAdmBanco: taxa de administracao do banco ou financeira}

\item{:}{txInflacaoIPCA: taxa IGPM para periodo}
}
\description{
Funcao para calcular previsão de rentabilidade de titulo NTN-B Principal
}
\author{
Rodrigo Almeida - rodrigo.almeida@gmail.com
}

