% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRentabilidadeNTNC.R
\name{getRentabilidadeNTNC}
\alias{getRentabilidadeNTNC}
\title{get Valor Rentabilidade Tesouro IGPM com Juros Semestrais (NTN-C)}
\usage{
getRentabilidadeNTNC(dataCompra, tituloPublico, dataVencimento, valorInvestido,
  txAdmBanco, txInflacaoIGPM)
}
\arguments{
\item{:}{dataCompra: Data da Compra do titulo no formato YYYY-MM-DD}

\item{:}{tituloPublico: Nome do Titulo Publico}

\item{:}{dataVencimento: Data de Vencimento do titulo no formato YYYY-MM-DD}

\item{:}{valorInvestido: Valor total investido no periodo}

\item{:}{txAdmBanco: taxa de administracao do banco ou financeira}

\item{:}{txInflacaoIGPM: taxa IGPM para periodo}
}
\description{
Funcao para calcular previsão de rentabilidade de titulo NTN-C
}
\author{
Rodrigo Almeida - rodrigo.almeida@gmail.com
}

