% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxaCompra.R
\name{getTaxaCompra}
\alias{getTaxaCompra}
\title{get Taxa do Papel no dia da Compra}
\usage{
getTaxaCompra(dataCompra, tituloPublico, dataVencimento)
}
\arguments{
\item{:}{dataCompra: Data da Compra do titulo no formato DD/MM/YYYY}

\item{:}{tituloPublico}

\item{:}{dataVencimento: Data de Vencimento no formato DD/MM/YYYY}
}
\description{
Funcao para recuperar a taxa do papel no dia da Compra
}
\author{
Rodrigo Almeida (rodrigo.almeida@gmail.com)
}

