% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTitulosPublicos.R
\name{getTitulosPublicos}
\alias{getTitulosPublicos}
\title{getTitulosPublicos}
\usage{
retorna nome dos titulos para Consulta
}
\arguments{
\item{dadosTesouro:}{Dados do tesouro lidos a partir da planilha csv 'precotaxatesourodireto.csv'}
}
\description{
os dados vem da seguinte url: http://www.tesourotransparente.gov.br/ckan/dataset/taxas-dos-titulos-ofertados-pelo-tesouro-direto

Retorna data.frame com as datas de vencimento dos papeis LFT, LTN, NTN-F, NTN-B Principal, NTN-B, NTN-C
LFT: Tesouro Selic
LTN: Tesouro Prefixado
NTN-F: Tesouro Prefixado com Juros Semestrais
NTN-B Principal: Tesouro IPCA
NTN-B: Tesouro IPCA com Juros Semestrais
NTN-C: Títulos vinculados ao IGPM
}
\author{
Rodrigo Almeida
}
\keyword{abertos,}
\keyword{dados}
\keyword{direto,}
\keyword{publicos}
\keyword{tesouro}
\keyword{titulos}

