% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValorCompra.R
\name{getValorCompra}
\alias{getValorCompra}
\title{get Valor compra Manhã}
\usage{
getValorCompra(dataCompra, tituloPublico, dataVencimento)
}
\arguments{
\item{:}{dataCompra: Data da Compra do titulo no formato YYYY-MM-DD}
}
\description{
Funcao para recuperar o valor da compra do titulo na data da compra
}
\author{
Rodrigo Almeida - rodrigo@lojadedados.com.br
}

