% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVencimentosTitulosPublicos.R
\name{getVencimentosTitulosPublicos}
\alias{getVencimentosTitulosPublicos}
\title{Prepara Dados}
\usage{
retorna planilha do tesouro direto
}
\description{
os dados vem da seguinte url: http://www.tesourotransparente.gov.br/ckan/dataset/taxas-dos-titulos-ofertados-pelo-tesouro-direto

Retorna data.frame com as datas de vencimento dos papeis LFT, LTN, NTN-F, NTN-B Principal, NTN-B, NTN-C
}
\author{
Rodrigo Almeida
}
\keyword{abertos}
\keyword{dados}
\keyword{direto,}
\keyword{tesouro}

