% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXTSTD.R
\name{getXTSTD}
\alias{getXTSTD}
\title{getXTSTD Recupera DADOS no formato XTS para visualização}
\usage{
retorna nome dos titulos para Consulta
}
\arguments{
\item{janelaLonga}{de avaliação da media movel exponencial}

\item{janelaCurta}{de avaliação da media movel exponencial}

\item{dataBase:}{Data Base do Titulo Publico (serie temporal)}

\item{infoApresentada:}{Um dos seguintes: Taxa.Compra.Manha, Taxa.Venda.Manha, PU.Compra.Manha, PU.Venda.Manha, PU.Base.Manha}
}
\author{
Rodrigo Almeida
}

