% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGrafico.R
\name{getGrafico}
\alias{getGrafico}
\title{getGrafico}
\usage{
getGrafico(tituloPublico, dataVencimento, janelaLonga, janelaCurta, tipoInfo)
}
\arguments{
\item{janelaLonga}{de avaliação da media movel exponencial}

\item{janelaCurta}{de avaliação da media movel exponencial}

\item{tituloPublico:}{Titulo Publico}

\item{dataVencimento:}{Data de vencimento do titulo}

\item{tipoInfo:}{Um dos seguintes: Taxa.Compra.Manha, Taxa.Venda.Manha, PU.Compra.Manha, PU.Venda.Manha, PU.Base.Manha}
}
\description{
Funcao para geracao de grafico do Tesouro
}
\author{
Rodrigo Almeida
}

