% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRentabilidadeNTNBPrincipal.R
\name{getRentabilidadeNTNBPrincipal}
\alias{getRentabilidadeNTNBPrincipal}
\title{get Valor Rentabilidade Tesouro IPCA  (NTN-B Principal)}
\usage{
getRentabilidadeNTNBPrincipal(dataCompra, tituloPublico = "Tesouro IPCA+",
  dataVencimento, valorInvestido, txAdmBanco = 0.1/100,
  txCustodia = 0.3/100, txInflacaoIPCA = 6/100)
}
\arguments{
\item{dataCompra}{Data da Compra do titulo no formato dd/mm/yyyy}

\item{tituloPublico}{Nome do Titulo Publico}

\item{dataVencimento}{Data de Vencimento do titulo no formato dd/mm/yyyy}

\item{valorInvestido}{Valor total investido no periodo}

\item{txAdmBanco}{taxa de administracao do banco ou financeira}

\item{txCustodia}{Taxa de Custodia (por padrao 0.3%)}

\item{txInflacaoIPCA}{taxa Inflacao para periodo}
}
\description{
Funcao para calcular previsão de rentabilidade de titulo NTN-B Principal
}
\author{
Rodrigo Almeida <rodrigo.almeida@gmail.com>
}

