% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYRange.R
\name{getYRange}
\alias{getYRange}
\title{getYRange define o intervalo do eixo Y que será visualizado no grafico}
\usage{
retorna intervalo que sera apresentado no eixo Y
}
\arguments{
\item{serieTemporal:}{Serie Temporal no formato XTS que sera plotada no grafico}
}
\value{
intervaloY: Intervalo
}
\author{
Rodrigo Almeida
}

