% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateData.R
\name{updateData}
\alias{updateData}
\title{update data}
\usage{
updateData()
}
\description{
Funcao para atualizar dados do tesouro direto que estao na pasta data com o nome 'precotaxatesourodireto.csv'
}
\author{
Rodrigo Almeida - rodrigo@lojadedados.com.br
}

